move vsphere-vclib vendor pkg to vsphere/lib directory
