+++ vsphere/lib/vsphere/vclib/BUILD
