+++ vsphere/lib/vsphere/vclib/connection.go
