+++ vsphere/lib/vsphere/vclib/constants.go
