+++ vsphere/lib/vsphere/vclib/custom_errors.go
