+++ vsphere/lib/vsphere/vclib/datastore.go
