+++ vsphere/lib/vsphere/vclib/diskmanagers/vmdm.go
@@ -22,10 +22,10 @@ import (
 	"hash/fnv"
 	"strings"
 
+	"github.com/libopenstorage/cloudops/vsphere/lib/vsphere/vclib"
 	"github.com/vmware/govmomi/object"
 	"github.com/vmware/govmomi/vim25/types"
 	"k8s.io/klog"
-	"k8s.io/kubernetes/pkg/cloudprovider/providers/vsphere/vclib"
 )
 
 // vmDiskManager implements VirtualDiskProvider interface for creating volume using Virtual Machine Reconfigure approach
