+++ vsphere/lib/vsphere/vclib/folder.go
