+++ vsphere/lib/vsphere/vclib/utils.go
