+++ vsphere/lib/vsphere/vclib/vmoptions.go
